Rename file: build-index.js -> entry.js
