+++ Nancy.ViewEngines.React/Nancy.ViewEngines.React.csproj
@@ -93,8 +93,8 @@
     <None Include="package.json" />
     <None Include="packages.config" />
     <None Include="targets\Nancy.ViewEngines.React.targets" />
-    <None Include="tools\gulp\build-index.js" />
     <None Include="tools\gulp\console.js" />
+    <None Include="tools\gulp\entry.js" />
     <None Include="tools\gulp\html.jsx" />
     <None Include="tools\gulp\layout.jsx" />
     <None Include="tools\gulp\options.js" />
