+++ Nancy.ViewEngines.React/tools/gulp/entry.js
